Principal Propagation (#89)
A fully augmented service!  Adds in a client for other services to
consume (e.g. you may want to provision Slurm or something but not
expose the full compute cluster to the end user), that handles token
injection and allows principal injection.

The API handler picks up the principal aware generators, that add the
correct labels to Kubernetes resources.  It also does an indirection
when setting up quota allocations to create/update/delete them from the
principal, not the faux Slurm controller.

In the controller proper, we can then read the principal labels (if
defined) and inject them into the HTTP requests to the region service so
it has an idea of who to charge for said resource utilization.