Server Event Listener
Rather than having a health monitor, that causes a potential split brain
problem, we should be more MQ and listen directly to server events.
This allows the main reconciler to respond to health status updates etc.
all in one place.  As an added bonus, we see the events when a server is
deleted out-of-band and schedule a replacement...