Switch to a New Compute Algorithm
Previously, we tried to be a bit like Kubernetes stateful sets, trying
to preserve an ordinal based hostname.  While this was great for Slurm
administrators, for evryone else it sucked, limiting what could be done.
So we return to the original plan of random hostnames, to stop people
making silly assumptions, and allow clever things like named evicitions
without leaving gaps that would have broken the orginal naming scheme.