Server Eviction
Rather than booting out a machine and having it come back to life, what
an end user may really want to do is to boot the machine out and have
the cluster scale down.  This performs that task and handles all the
potential races, orchestration between the cluster, servers and quota
allocations.