Smorgasbord of Fixes (#121)
Cluster deletion doesn't ignore an already deleting cluster so it will
result in an error as allocations etc no longer exist.

Merges in a new core package that deletes resources regardless of the
paused flag, which is an admin only hack anyway.

Rethinks the eviction stuff by siply scaling down the cluster and
annotating it with the nodes that should be deleted.  The reconciler
will then pick up this information and do as is necessary without some
vast racy interplay of pausing, updating, unpausing.  We add a load of
API protections in because it was still possible to repeatedly evict and
eventiually scale the cluster to zero, or worse still a massive number!