Fix Nil Dereference
The security group rule definition changed recently to make the port
definition optional (e.g. if not specified then match all ports for L4
traffic).  Unfortunately this didn't trigger a warning of any kind when
generating the request in the cluster controller and we were writing to
a nil pointer struct.  As ports are required for this (soon to be) old
API, just add in a blank default value and fill it in.